Change the behavior of genre buttons
